Extract `NullRelationTest` from `RelationTest`
`test/cases/relations_test.rb` file has too much lines (2000 over).
So I extracted `NullRelationTest` to the dedicated file.
